delete add_metadata folder and update it to user_metadata_function. also update the readme to documentation.
